Fix name collision issue
This name collision resolves the issue, in which the reference
to the original "class" function of Scrap Mechanic, was changed for
all tools and other mods too.

With this update it is again possible to have multiple scripted parts
in the world, which have been written using this library
and to load those without any issue. Other mods should be compatible
with mods using this library.